# Speed up prepare_tokenizer_data on large treebanks