Avoid excessive copying of the full-text on each paragraph check.

`stanza.utils.datasets.prepare_tokenizer_treebank` took around 2 hours on my custom treebank (~60 MiB .txt file). 
After this fix it finished in 2 minutes.  

Results for other large UD treebanks:
|                      | before | after |
|----------------------|--------|---------|
| UD_Russian-SynTagRus | 46s    | 23s     |
| UD_German-HDT        | 6m 12s | 50s     |